Use eslint-config-standard-with-typescript
