+++ .eslintrc.js
@@ -6,21 +6,45 @@ module.exports = {
     webextensions: true
   },
   parser: '@typescript-eslint/parser',
-  plugins: [
-    '@typescript-eslint'
-  ],
-  extends: [
-    'semistandard',
-    'plugin:@typescript-eslint/recommended'
-  ],
+  extends: 'standard-with-typescript',
   parserOptions: {
     sourceType: 'module'
   },
+  plugins: [
+    'svelte3',
+    '@typescript-eslint'
+  ],
+  overrides: [
+    {
+      files: ['*.svelte'],
+      processor: 'svelte3/svelte3',
+      rules: {
+        'import/first': 'off',
+        'import/no-duplicates': 'off',
+        'import/no-mutable-exports': 'off',
+        'no-multiple-empty-lines': [
+          'error',
+          { max: 2 }
+        ]
+      }
+    }
+  ],
   rules: {
     'linebreak-style': [
       'error',
       'unix'
-    ]
+    ],
+    semi: [
+      'error',
+      'always'
+    ],
+    'no-extra-semi': 'error'
   },
-  root: true
+  settings: {
+    'svelte3/typescript': () => require('typescript')
+  },
+  globals: {
+    Ytc: 'readonly',
+    Chat: 'readonly'
+  }
 };
