+++ package.json
@@ -11,15 +11,15 @@
   "dependencies": {
     "@mdi/js": "^5.9.55",
     "postcss-input-range": "^4.0.0",
-    "postcss-url": "^8.0.0",
+    "postcss-url": "^10.1.3",
     "smelte": "^1.1.6",
-    "tailwindcss-elevation": "^0.3.1"
+    "tailwindcss-elevation": "^1.0.1"
   },
   "devDependencies": {
     "@fullhuman/postcss-purgecss": "^4.0.3",
     "@tsconfig/svelte": "^2.0.1",
     "@types/chrome": "^0.0.149",
-    "@typescript-eslint/eslint-plugin": "^4.28.5",
+    "@typescript-eslint/eslint-plugin": "4",
     "@typescript-eslint/parser": "^4.28.5",
     "autoprefixer": "^10.3.1",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
@@ -27,12 +27,11 @@
     "cross-env": "^7.0.3",
     "css-loader": "^6.2.0",
     "cssnano": "^5.0.7",
-    "eslint": "^7.31.0",
-    "eslint-config-semistandard": "^16.0.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
+    "eslint": "7",
+    "eslint-config-standard-with-typescript": "^20.0.0",
+    "eslint-plugin-import": "2",
+    "eslint-plugin-node": "11",
+    "eslint-plugin-promise": "4",
     "eslint-plugin-svelte3": "^3.2.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
